requirements.txt
skyfield
numpy
requests
json
time
astropy
os
python-dotenv
matplotlib